# docs/swagger.yaml
definitions:
  domain.Bus:
    properties:
      id:
        type: string
      number:
        type: string
      total_seats:
        type: integer
    type: object
...
  errs.Err:
    properties:
      errDesc:
        type: string
      error:
        type: string
    type: object
...
  model.CreateRequest:
    properties:
      number:
        type: string
      total_seats:
        type: integer
    type: object
...
  model.ForgotPasswordRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
...
  model.SignupRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  model.SignupResponse:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/model.UserResponse'
    type: object
...
  model.UserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      firstname:
        type: string
      id:
        type: string
      lastname:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
  model.VerifyResetCodeRequest:
    properties:
      code:
        type: string
      email:
        type: string
      new_password:
        type: string
    required:
    - code
    - email
    - new_password
    type: object
info:
  contact: {}
  description: API documentation for Aulway.
  title: Aulway API
  version: "1.0"
paths:
...
  /api/buses:
    get:
      consumes:
      - application/json
      description: Retrieve a list of buses
      parameters:
      - description: 'Page number for pagination (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Page size for pagination (default: 30)'
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of buses
          schema:
            items:
              items:
                $ref: '#/definitions/domain.Bus'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      security:
      - BearerAuth: []
      summary: Get Buses List
      tags:
      - bus
    post:
      consumes:
      - application/json
      description: Create Bus
      parameters:
      - description: Request Body
        in: body
        name: requestBody
        required: true
        schema:
          $ref: '#/definitions/model.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/domain.Bus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      security:
      - BearerAuth: []
      tags:
      - bus
  /api/buses/{busId}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Bus ID
        in: path
        name: busId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: bus deleted
          schema:
            items:
              type: string
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      security:
      - BearerAuth: []
      summary: Delete bus by id
      tags:
      - bus
    get:
      consumes:
      - application/json
      description: Retrieve a bus by its ID
      parameters:
      - description: Bus ID
        in: path
        name: busId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/domain.Bus'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      security:
      - BearerAuth: []
      summary: Get Bus
      tags:
      - bus
...
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: forgot password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid request body
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: Forgot password
      tags:
      - auth
  /auth/forgot-password/verify:
    post:
      consumes:
      - application/json
      parameters:
      - description: required
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.VerifyResetCodeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request - Invalid request body
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: verify forgot password
      tags:
      - auth
...
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Register a new user and return an access token.
      parameters:
      - description: Signup Request Body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.SignupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.SignupResponse'
        "400":
          description: Bad Request - Invalid request body
          schema:
            $ref: '#/definitions/errs.Err'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.Err'
      summary: User Signup
      tags:
      - auth
securityDefinitions:
  BearerAuth:
    description: Use "Bearer {your-firebase-token}"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"